Test generalError middleware
