--- src/server/middlewares/errorMiddlewares/error.Middlewares.test.ts
@@ -1,21 +1,0 @@
-import { type NextFunction, type Request, type Response } from "express";
-import { notFoundError } from "./errorMiddlewares";
-
-const req = {} as Request;
-const res = {
-  status: jest.fn().mockReturnThis(),
-  json: jest.fn(),
-} as Partial<Response>;
-const next = jest.fn() as NextFunction;
-
-beforeEach(() => jest.clearAllMocks());
-
-describe("Given a notFoundError middleware", () => {
-  describe("When it receives a response", () => {
-    test("Then it should call its next method", () => {
-      notFoundError(req, res as Response, next);
-
-      expect(next).toHaveBeenCalled();
-    });
-  });
-});
